backend/requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
demucs==4.0.1
dora_search==0.1.12
einops==0.8.1
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
julius==0.2.7
lameenc==1.8.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
omegaconf==2.3.0
openunmix==1.3.0
pip==25.0.1
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
retrying==1.3.4
setuptools==80.3.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
submitit==1.5.2
sympy==1.14.0
syncedlyrics==1.0.1
torch==2.7.0
torchaudio==2.7.0
tqdm==4.67.1
treetable==0.2.5
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
yt-dlp==2025.4.30
